Remove call to setlocale (closes #5538):
Turns out the environment variable alone is sufficient. I'd previously thought it wasn't, but that was actually because I'd been setting it to "C.UTF-8", which Crystax apparently doesn't recognize, rather than "en_US.UTF-8".

The setlocale call had no effect on the Crystax locale system because the Chaquopy runtime is compiled with the Google NDK, not Crystax. On API 26, the Google implementation of setlocale will accept and return a small set of values, including both of the above and an empty string. But on API 19, it always returns NULL, which was causing us to crash.