Make Gradle plugin use correct extension for bootstrap ZIP:
The build process for version 8.0.0 generated a file with the extension .imy, but the Gradle
plugin still used the extension .zip. I went through the whole release process without noticing
this, because there was still a leftover .zip file in the Maven repository. Fortunately there
had been no changes to the file's content since it was renamed.