Apply pre-API-level-23 dynamic linker workarounds to base ExtensionFileLoader class:
In this case, the issue was that _csv.so was unnecessarily included in the bootstrap modules list. When it was later imported through the standard ExtensionFileLoader, the dynamic linker's LD_LIBRARY_PATH was not fixed, which caused the "libpython3.X.so not found" error from #5563.

However, a similar problem could also occur if a module in bootstrap-native is needed on some Python versions but not others, which may already be the case, for all I know.